+++ src/App.css
@@ -1,0 +1,346 @@
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@600&display=swap');
+
+body {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+body,
+html {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  height: 100%;
+  background-color: #eceef2;
+}
+
+/* HEADER */
+header {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  height: auto;
+  box-shadow: inset 0px -1px #aaaaaa;
+  font-family: 'Inter', sans-serif;
+  background-color: #eceef2;
+}
+
+.header-text {
+  padding-left: 15%;
+  margin: 15px 0;
+}
+
+.header-text-1 {
+  margin: 0;
+  font-size: 18px;
+  line-height: 24px;
+  color: #111111;
+}
+
+.header-text-2 {
+  margin: 0;
+  font-weight: bold;
+  font-size: 40px;
+  line-height: 48px;
+  color: #074ee8;
+}
+
+/* MAIN */
+
+main {
+  height: auto;
+  background-color: #eceef2;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.form-container {
+  width: 650px;
+  height: auto;
+  background: #ffffff;
+  border-radius: 20px;
+  max-width: 85%;
+  margin: 1% 0;
+}
+
+.tab-container {
+  margin: 1% 5%;
+}
+
+/* ALL TABS */
+
+.title,
+.input-block,
+.tabs-container {
+  margin: 20px 0;
+}
+
+#third-tab,
+#second-tab,
+#first-tab {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+}
+
+.btn-to-right {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.btn-space-between {
+  display: flex;
+  justify-content: space-between;
+  /* align-items: baseline; */
+}
+
+.hide {
+  display: none !important;
+}
+
+button {
+  cursor: pointer;
+}
+
+/* To make responsive */
+
+@media (max-width: 600px) {
+  .tab-container {
+    margin: 0;
+  }
+
+  .vsc-initialized,
+  body,
+  html,
+  main {
+    background-color: #ffffff !important;
+  }
+
+  .header-text {
+    margin: 15px 5px;
+    padding-left: 5%;
+  }
+
+  textarea:focus,
+  input:focus:not(.input-heart),
+  select:focus {
+    box-shadow: 0 0 0 0;
+    border: 1.8px solid #aaaaaa;
+    outline: 0;
+  }
+
+  .select {
+    padding: 11px 12px;
+  }
+
+  .input-heart {
+    font-size: 12px;
+    padding: 11px 12px;
+  }
+
+  .certificates-opt {
+    margin: 0;
+    max-width: 150px;
+  }
+
+  .certificates-added li {
+    margin: 0;
+    width: 168px;
+  }
+}
+
+@media (max-width: 460px) {
+  .flex-line {
+    flex-direction: row;
+    flex-wrap: wrap;
+  }
+
+  .medium-input,
+  .small-input {
+    width: 100% !important;
+  }
+
+  .small-input {
+    margin: 20px 0;
+  }
+
+  .line-title {
+    margin: 0 !important;
+  }
+
+  .four-inputs-by-line {
+    width: 45% !important;
+    margin: 10px 0;
+  }
+}
+
+@media (max-width: 342px) {
+  .input-block.btn-space-between.btn-more {
+    flex-wrap: wrap;
+  }
+
+  .certificates-list {
+    margin-right: 2px;
+  }
+}
+
+@media (min-width: 2000px) {
+  body,
+  html {
+    width: auto;
+    height: auto;
+  }
+
+  header {
+    height: 290px;
+  }
+
+  .header-text {
+    margin: 15px 5px;
+    padding-left: 13%;
+  }
+
+  .header-text-1 {
+    font-size: 40px;
+    line-height: 50px;
+    margin-bottom: 20px;
+  }
+
+  .header-text-2 {
+    font-size: 82px;
+    line-height: 86px;
+  }
+
+  .form-container {
+    width: 60%;
+    height: 70%;
+  }
+
+  textarea:focus,
+  input:focus:not(.input-heart),
+  select:focus {
+    box-shadow: 0 0 0 0;
+    border: 3px solid #aaaaaa;
+    outline: 0;
+  }
+
+  .inputs,
+  .select,
+  #input_btn {
+    font-size: 36px;
+    border-width: 3px;
+  }
+
+  .labels {
+    font-size: 40px;
+    line-height: 48px;
+  }
+
+  .title {
+    font-size: 40px;
+    line-height: 48px;
+  }
+
+  .title,
+  .input-block,
+  .tabs-container {
+    margin: 4% 0 !important;
+  }
+
+  .tabs {
+    height: 100px;
+    box-shadow: none;
+    border-bottom: 5px solid #aaaaaa;
+    font-family: 'Nunito', sans-serif;
+    font-size: 40px !important;
+    line-height: 48px !important;
+    cursor: pointer;
+  }
+
+  .tabs-container .active {
+    box-shadow: none;
+    color: #074ee8;
+    border-bottom: 5px solid #074ee8;
+  }
+
+  .line-title,
+  .check-terms {
+    font-size: 40px !important;
+    line-height: 48px !important;
+  }
+
+  .checkmark {
+    height: 40px;
+    width: 40px;
+    border-width: 3px;
+  }
+
+  .check-terms {
+    padding-left: 65px;
+  }
+
+  .check-terms .checkmark:after {
+    left: 12px;
+    top: 2px;
+    width: 10px;
+    height: 25px;
+    border-width: 0 4px 4px 0;
+  }
+
+  .select {
+    padding: 11px 12px;
+  }
+
+  .input-heart,
+  .certificates-list {
+    font-size: 36px;
+    padding: 11px 12px;
+  }
+
+  .certificates-list ul {
+    margin-top: 8px;
+  }
+
+  .certificates-opt {
+    margin: 0;
+    width: 600px;
+    padding: 20px;
+    box-sizing: border-box;
+    border-radius: 12px;
+  }
+
+  .certificates-border {
+    border-radius: 12px 12px 0px 0px;
+  }
+
+  svg {
+    width: 40px;
+    height: 40px !important;
+  }
+
+  .certificates-added li {
+    margin: 0;
+    padding: 0px 10px;
+    box-sizing: border-box;
+    width: 600px;
+  }
+
+  .btn {
+    border-radius: 12px;
+    padding: 30px;
+    font-size: 36px;
+    font-weight: normal;
+  }
+
+  .btn * {
+    padding: 0 10px;
+  }
+
+  .showError,
+  .message-error {
+    font-size: 36px !important;
+    line-height: 38px !important;
+  }
+}
