--- src/ViewTabs/BodyStructure/BodyStructure.css
@@ -1,105 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@600&display=swap');
-
-body {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-body,
-html {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  height: 100%;
-  background-color: #eceef2;
-}
-
-/* HEADER */
-header {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  height: auto;
-  box-shadow: inset 0px -1px #aaaaaa;
-  font-family: 'Inter', sans-serif;
-  background-color: #eceef2;
-}
-
-.header-text {
-  padding-left: 15%;
-  margin: 15px 0;
-}
-
-.header-text-1 {
-  margin: 0;
-  font-size: 18px;
-  line-height: 24px;
-  color: #111111;
-}
-
-.header-text-2 {
-  margin: 0;
-  font-weight: bold;
-  font-size: 40px;
-  line-height: 48px;
-  color: #074ee8;
-}
-
-/* MAIN */
-
-main {
-  height: auto;
-  background-color: #eceef2;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.form-container {
-  width: 650px;
-  height: auto;
-  background: #ffffff;
-  border-radius: 20px;
-  max-width: 85%;
-  margin: 1% 0;
-}
-
-.tab-container {
-  margin: 1% 5%;
-}
-
-/* ALL TABS */
-
-.title,
-.input-block,
-.tabs-container {
-  margin: 20px 0;
-}
-
-#third-tab,
-#second-tab,
-#first-tab {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-}
-
-.btn-to-right {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.btn-space-between {
-  display: flex;
-  justify-content: space-between;
-  /* align-items: baseline; */
-}
-
-.hide {
-  display: none !important;
-}
-
-button {
-  cursor: pointer;
-}
